requirements.txt
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
click==8.1.8
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
googletrans==4.0.2
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
numpy==2.2.4
packaging==24.2
pandas==2.2.3
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
uvicorn==0.34.0
